+++ include/JSystem/JStudio/JStudio_JStage/object-actor.h
@@ -1,6 +1,6 @@
 #ifndef OBJECT_ACTOR_H
 #define OBJECT_ACTOR_H
 
-#include "JSystem/JStudio/JStudio_JStage/control.h"
+#include "JSystem/JStudio/JStudio_JStage/control.h" // IWYU pragma: export
 
 #endif /* OBJECT_ACTOR_H */
