+++ src/JSystem/JStudio/JStudio_JStage/object-actor.cpp
@@ -4,260 +4,20 @@
 //
 
 #include "JSystem/JStudio/JStudio_JStage/object-actor.h"
-#include "dol2asm.h"
-
-//
-// Forward References:
-//
-
-extern "C" void __ct__Q214JStudio_JStage14TAdaptor_actorFPCQ26JStage7TSystemPQ26JStage6TActor();
-extern "C" void __dt__Q214JStudio_JStage14TAdaptor_actorFv();
-extern "C" void adaptor_do_prepare__Q214JStudio_JStage14TAdaptor_actorFv();
-extern "C" void adaptor_do_begin__Q214JStudio_JStage14TAdaptor_actorFv();
-extern "C" void adaptor_do_end__Q214JStudio_JStage14TAdaptor_actorFv();
-extern "C" void adaptor_do_update__Q214JStudio_JStage14TAdaptor_actorFUl();
-extern "C" void adaptor_do_data__Q214JStudio_JStage14TAdaptor_actorFPCvUlPCvUl();
-extern "C" void
-adaptor_do_SHAPE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_ANIMATION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_ANIMATION_MODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_TEXTURE_ANIMATION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_TEXTURE_ANIMATION_MODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_PARENT__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_PARENT_NODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_PARENT_ENABLE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_PARENT_FUNCTION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_RELATION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_RELATION_NODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-adaptor_do_RELATION_ENABLE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void
-setJSG_ID___Q214JStudio_JStage14TAdaptor_actorFMQ26JStage6TActorFPCvPvUl_vQ37JStudio4data15TEOperationDataPCvUl();
-extern "C" void setJSG_SRT___Q214JStudio_JStage14TAdaptor_actorFPCQ27JStudio8TControl();
-extern "C" void getJSG_SRT___Q214JStudio_JStage14TAdaptor_actorFPCQ27JStudio8TControl();
-extern "C" void
-__cl__Q314JStudio_JStage14TAdaptor_actor26TVVOutput_ANIMATION_FRAME_CFfPQ27JStudio8TAdaptor();
-extern "C" void __dt__Q314JStudio_JStage14TAdaptor_actor26TVVOutput_ANIMATION_FRAME_Fv();
-extern "C" void func_8028B198();
-extern "C" void __ct__Q27JStudio14TVariableValueFv();
-extern "C" void func_8028B580(void* _this);
-extern "C" void func_8028B5E0(void* _this, f32, JStudio::TAdaptor*);
-extern "C" u8 saoVVOutput___Q214JStudio_JStage14TAdaptor_actor[64];
-extern "C" u8
-    saoVVOutput_ANIMATION_FRAME___Q214JStudio_JStage14TAdaptor_actor[144 + 4 /* padding */];
-
-//
-// External References:
-//
-
-extern "C" void toFunction_outside__Q27JStudio14TFunctionValueFi();
-extern "C" void getFromTransformation_SRxyzT__Q27JStudio4mathFP3VecP3VecP3VecPA4_Cf();
-extern "C" void __dt__Q37JStudio14TVariableValue7TOutputFv();
-extern "C" void update_immediate___Q27JStudio14TVariableValueFPQ27JStudio14TVariableValued();
-extern "C" void adaptor_setVariableValue_Vec__Q27JStudio8TAdaptorFPCUlRC3Vec();
-extern "C" void adaptor_getVariableValue_Vec__Q27JStudio8TAdaptorCFP3VecPCUl();
-extern "C" void __dt__Q27JStudio14TAdaptor_actorFv();
-extern "C" void
-transform_toGlobalFromLocal__14JStudio_JStageFPA4_fRCQ37JStudio8TControl39TTransform_translation_rotation_scalingPCQ26JStage7TObjectUl();
-extern "C" void
-transform_toLocalFromGlobal__14JStudio_JStageFPA4_fRCQ37JStudio8TControl39TTransform_translation_rotation_scalingPCQ26JStage7TObjectUl();
-extern "C" void adaptor_object_data___Q214JStudio_JStage16TAdaptor_object_FPCvUlPCvUl();
-extern "C" void adaptor_object_findJSGObject___Q214JStudio_JStage16TAdaptor_object_FPCc();
-extern "C" void
-adaptor_object_findJSGObjectNode___Q214JStudio_JStage16TAdaptor_object_FPCQ26JStage7TObjectPCc();
-extern "C" void __dl__FPv();
-extern "C" void __register_global_object();
-extern "C" void __construct_array();
-extern "C" void __ptmf_scall();
-extern "C" void _savegpr_27();
-extern "C" void _savegpr_28();
-extern "C" void _savegpr_29();
-extern "C" void _restgpr_27();
-extern "C" void _restgpr_28();
-extern "C" void _restgpr_29();
-extern "C" u8 const sauVariableValue_3_TRANSLATION_XYZ__Q27JStudio14TAdaptor_actor[12];
-extern "C" u8 const sauVariableValue_3_ROTATION_XYZ__Q27JStudio14TAdaptor_actor[12];
-extern "C" u8 const sauVariableValue_3_SCALING_XYZ__Q27JStudio14TAdaptor_actor[12];
-extern "C" extern u8 const __ptmf_null[12 + 4 /* padding */];
-extern "C" extern void* __vt__Q27JStudio14TAdaptor_actor[20];
-extern "C" extern void* __vt__Q27JStudio8TAdaptor[8];
-extern "C" extern void* __vt__Q37JStudio14TVariableValue7TOutput[4];
-extern "C" u8 soOutput_none___Q27JStudio14TVariableValue[4 + 4 /* padding */];
-
-//
-// Declarations:
-//
-
-/* ############################################################################################## */
-/* 803C5730-803C573C 022850 000C+00 2/2 0/0 0/0 .data            @890 */
-SECTION_DATA static u8 lit_890[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x58, 0x00, 0x00, 0x00, 0x00,
-};
-
-/* 803C573C-803C5748 02285C 000C+00 1/1 0/0 0/0 .data            @899 */
-SECTION_DATA static u8 lit_899[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00, 0x00,
-};
-
-/* 803C5748-803C5754 022868 000C+00 1/1 0/0 0/0 .data            @916 */
-SECTION_DATA static u8 lit_916[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x7C, 0x00, 0x00, 0x00, 0x00,
-};
-
-/* 803C5754-803C5760 022874 000C+00 0/1 0/0 0/0 .data            @1083 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1083[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x74, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C5760-803C576C 022880 000C+00 0/1 0/0 0/0 .data            @1084 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1084[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C576C-803C5778 02288C 000C+00 0/1 0/0 0/0 .data            @1089 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1089[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x68, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C5778-803C5784 022898 000C+00 0/1 0/0 0/0 .data            @1090 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1090[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x64, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C5784-803C5790 0228A4 000C+00 0/1 0/0 0/0 .data            @1091 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1091[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x6C, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C5790-803C579C 0228B0 000C+00 0/1 0/0 0/0 .data            @1096 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1096[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x84, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C579C-803C57A8 0228BC 000C+00 0/1 0/0 0/0 .data            @1097 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1097[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C57A8-803C57B4 0228C8 000C+00 0/1 0/0 0/0 .data            @1098 */
-#pragma push
-#pragma force_active on
-SECTION_DATA static u8 lit_1098[12] = {
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x88, 0x00, 0x00, 0x00, 0x00,
-};
-#pragma pop
-
-/* 803C57B4-803C57C4 0228D4 0010+00 1/2 0/0 0/0 .data
- * __vt__Q314JStudio_JStage14TAdaptor_actor26TVVOutput_ANIMATION_FRAME_ */
-SECTION_DATA extern void* __vt__Q314JStudio_JStage14TAdaptor_actor26TVVOutput_ANIMATION_FRAME_[4] = {
-    (void*)NULL /* RTTI */,
-    (void*)NULL,
-    (void*)
-        __cl__Q314JStudio_JStage14TAdaptor_actor26TVVOutput_ANIMATION_FRAME_CFfPQ27JStudio8TAdaptor,
-    (void*)__dt__Q314JStudio_JStage14TAdaptor_actor26TVVOutput_ANIMATION_FRAME_Fv,
-};
-
-/* 803C57C4-803C5864 0228E4 0050+50 2/2 0/0 0/0 .data __vt__Q214JStudio_JStage14TAdaptor_actor */
-SECTION_DATA extern void* __vt__Q214JStudio_JStage14TAdaptor_actor[20 + 20 /* padding */] = {
-    (void*)NULL /* RTTI */,
-    (void*)NULL,
-    (void*)__dt__Q214JStudio_JStage14TAdaptor_actorFv,
-    (void*)adaptor_do_prepare__Q214JStudio_JStage14TAdaptor_actorFv,
-    (void*)adaptor_do_begin__Q214JStudio_JStage14TAdaptor_actorFv,
-    (void*)adaptor_do_end__Q214JStudio_JStage14TAdaptor_actorFv,
-    (void*)adaptor_do_update__Q214JStudio_JStage14TAdaptor_actorFUl,
-    (void*)adaptor_do_data__Q214JStudio_JStage14TAdaptor_actorFPCvUlPCvUl,
-    (void*)
-        adaptor_do_PARENT__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_PARENT_NODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_PARENT_ENABLE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_PARENT_FUNCTION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_RELATION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_RELATION_NODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_RELATION_ENABLE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_SHAPE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_ANIMATION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_ANIMATION_MODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_TEXTURE_ANIMATION__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    (void*)
-        adaptor_do_TEXTURE_ANIMATION_MODE__Q214JStudio_JStage14TAdaptor_actorFQ37JStudio4data15TEOperationDataPCvUl,
-    /* padding */
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-};
 
 /* 8028A5F0-8028A6B4 284F30 00C4+00 0/0 1/1 0/0 .text
  * __ct__Q214JStudio_JStage14TAdaptor_actorFPCQ26JStage7TSystemPQ26JStage6TActor */
 JStudio_JStage::TAdaptor_actor::TAdaptor_actor(JStage::TSystem const* param_1,
                                                    JStage::TActor* param_2)         :
                                                    TAdaptor_object_(param_1,param_2) {
-  field_0x130 = 0;
-  field_0x134 = 0;
-  field_0x138 = 0;
-  field_0x13c = NULL;
-  field_0x140 = -1;
-  field_0x144 = 0;
-  field_0x148 = NULL;
-  field_0x14c = -1;
+    field_0x130 = 0;
+    field_0x134 = 0;
+    field_0x138 = 0;
+    field_0x13c = NULL;
+    field_0x140 = -1;
+    field_0x144 = 0;
+    field_0x148 = NULL;
+    field_0x14c = -1;
 }
 
 /* 8028A6B4-8028A728 284FF4 0074+00 1/0 0/0 0/0 .text __dt__Q214JStudio_JStage14TAdaptor_actorFv
@@ -267,52 +27,16 @@ JStudio_JStage::TAdaptor_actor::~TAdaptor_actor() {
 }
 
 
-/* ############################################################################################## */
-/* 80431188-80431194 05DEA8 000C+00 1/1 0/0 0/0 .bss             @1087 */
-static u8 lit_1087[12];
-
-/* 80431194-804311A0 05DEB4 000C+00 0/1 0/0 0/0 .bss             @1088 */
-#pragma push
-#pragma force_active on
-static u8 lit_1088[12];
-#pragma pop
-
-/* 804311A0-804311E0 05DEC0 0040+00 2/3 0/0 0/0 .bss
- * saoVVOutput___Q214JStudio_JStage14TAdaptor_actor             */
-JStudio_JStage::TAdaptor_actor::TVVOutputObject JStudio_JStage::TAdaptor_actor::saoVVOutput_[2];
-
-/* 804311E0-804311EC 05DF00 000C+00 0/1 0/0 0/0 .bss             @1095 */
-#pragma push
-#pragma force_active on
-static u8 lit_1095[12];
-#pragma pop
-
-/* 804311EC-804311F8 05DF0C 000C+00 0/1 0/0 0/0 .bss             @1102 */
-#pragma push
-#pragma force_active on
-static u8 lit_1102[12];
-#pragma pop
-
-/* 804311F8-80431204 05DF18 000C+00 0/1 0/0 0/0 .bss             @1103 */
-#pragma push
-#pragma force_active on
-static u8 lit_1103[12];
-#pragma pop
-
-/* 80431204-80431298 05DF24 0090+04 2/3 0/0 0/0 .bss
- * saoVVOutput_ANIMATION_FRAME___Q214JStudio_JStage14TAdaptor_actor */
-JStudio_JStage::TAdaptor_actor::TVVOutput_ANIMATION_FRAME_ JStudio_JStage::TAdaptor_actor::saoVVOutput_ANIMATION_FRAME_[3];
-
 /* 8028A728-8028A7B4 285068 008C+00 1/0 0/0 0/0 .text
  * adaptor_do_prepare__Q214JStudio_JStage14TAdaptor_actorFv     */
 void JStudio_JStage::TAdaptor_actor::adaptor_do_prepare() {
-    TVVOutputObject* pOutput = saoVVOutput_;
+    const TVVOutputObject* pOutput = saoVVOutput_;
     while (!pOutput->isEnd_()) {
         pOutput->adaptor_setOutput_(this);
         pOutput++;
     }
 
-    TVVOutput_ANIMATION_FRAME_* pAnimOutput = saoVVOutput_ANIMATION_FRAME_;
+    const TVVOutput_ANIMATION_FRAME_* pAnimOutput = saoVVOutput_ANIMATION_FRAME_;
     while (!pAnimOutput->isEnd_()) {
         pAnimOutput->adaptor_setOutput_(this);
         pAnimOutput++;
@@ -322,23 +46,29 @@ void JStudio_JStage::TAdaptor_actor::adaptor_do_prepare() {
 
 /* 8028A7B4-8028A8D8 2850F4 0124+00 1/0 0/0 0/0 .text
  * adaptor_do_begin__Q214JStudio_JStage14TAdaptor_actorFv       */
-// NONMATCHING - instruction order
 void JStudio_JStage::TAdaptor_actor::adaptor_do_begin() {
     adaptor_object_begin_();
-    JStage::TActor* pActor = get_pJSG_();
-    const JStudio::TObject* pObject = adaptor_getObject();
-    JUT_ASSERT(72, pObject != 0);
-    JStudio::TControl* pControl = pObject->getControl();
-    JUT_ASSERT(74, pControl != 0);
-    getJSG_SRT_(adaptor_getObject()->getControl());
+    const JStage::TActor* pActor = get_pJSG_();
+#ifdef  DEBUG
+    { // This block is needed to match the stack in debug
+        const JStudio::TObject* pObject = adaptor_getObject();
+        JUT_ASSERT(72, pObject!=0);
+        const JStudio::TControl* pControl = pObject->getControl();
+        JUT_ASSERT(74, pControl!=0);
+        getJSG_SRT_(pControl);
+    }
+#else
+    // Fakematch, load order is wrong in retail when using the inlines shown in debug
+    getJSG_SRT_((const JStudio::TControl*)(adaptor_getObject()->pControl));
+#endif
 
-    TVVOutputObject* pOutput = saoVVOutput_;
+    const TVVOutputObject* pOutput = saoVVOutput_;
     while (!pOutput->isEnd_()) {
         pOutput->setVariableValue_(pActor, this);
         pOutput++;
     }
 
-    TVVOutput_ANIMATION_FRAME_* pAnimOutput = saoVVOutput_ANIMATION_FRAME_;
+    const TVVOutput_ANIMATION_FRAME_* pAnimOutput = saoVVOutput_ANIMATION_FRAME_;
     while (!pAnimOutput->isEnd_()) {
         pAnimOutput->setVariableValue_(pActor, this);
         pAnimOutput++;
@@ -356,7 +86,7 @@ void JStudio_JStage::TAdaptor_actor::adaptor_do_end() {
 void JStudio_JStage::TAdaptor_actor::adaptor_do_update(u32 param_1) {
     const JStudio::TObject* pObject = adaptor_getObject();
     JUT_ASSERT(104, pObject != 0);
-    JStudio::TControl* pControl = pObject->getControl();
+    const JStudio::TControl* pControl = pObject->getControl();
     JUT_ASSERT(106, pControl != 0);
     setJSG_SRT_(pControl);
     pJSGObject_->JSGUpdate();
@@ -427,12 +157,12 @@ void JStudio_JStage::TAdaptor_actor::adaptor_do_TEXTURE_ANIMATION_MODE(
 void JStudio_JStage::TAdaptor_actor::adaptor_do_PARENT(
     JStudio::data::TEOperationData operation, void const* pContent, u32 uSize) {
     switch (operation) {
-	case JStudio::data::UNK_0x18:
+    case JStudio::data::UNK_0x18:
         JUT_ASSERT(203, pContent!=0);
         JUT_ASSERT(204, uSize!=0);
-		field_0x13c = adaptor_object_findJSGObject_((const char*)pContent);
-		break;
-	}
+        field_0x13c = adaptor_object_findJSGObject_((const char*)pContent);
+        break;
+    }
 }
 
 /* 8028AB0C-8028AB68 28544C 005C+00 1/0 0/0 0/0 .text
@@ -441,17 +171,17 @@ void JStudio_JStage::TAdaptor_actor::adaptor_do_PARENT(
 void JStudio_JStage::TAdaptor_actor::adaptor_do_PARENT_NODE(
     JStudio::data::TEOperationData operation, void const* pContent, u32 uSize) {
     switch (operation) {
-	case JStudio::data::UNK_0x18:
+    case JStudio::data::UNK_0x18:
         JUT_ASSERT(226, pContent!=0);
         JUT_ASSERT(227, uSize!=0);
-		field_0x140 = adaptor_object_findJSGObjectNode_(field_0x13c, (const char*)pContent);
-		break;
-	case JStudio::data::UNK_0x19:
+        field_0x140 = adaptor_object_findJSGObjectNode_(field_0x13c, (const char*)pContent);
+        break;
+    case JStudio::data::UNK_0x19:
         JUT_ASSERT(237, uSize==4);
         JUT_ASSERT(238, pContent!=0);
-		field_0x140 = *(u32*)pContent;
-		break;
-	}
+        field_0x140 = *(u32*)pContent;
+        break;
+    }
 }
 
 /* 8028AB68-8028ABF0 2854A8 0088+00 1/0 0/0 0/0 .text
@@ -460,28 +190,28 @@ void JStudio_JStage::TAdaptor_actor::adaptor_do_PARENT_NODE(
 void JStudio_JStage::TAdaptor_actor::adaptor_do_PARENT_ENABLE(
     JStudio::data::TEOperationData operation, void const* pContent, u32 uSize) {
     switch (operation) {
-	case JStudio::data::UNK_0x2:
+    case JStudio::data::UNK_0x2:
         JUT_ASSERT(257, uSize==4);
         JUT_ASSERT(258, pContent!=0);
-		bool v0 = (*(u32*)pContent != 0);
-		if (field_0x144 != v0) {
-			field_0x144 = v0;
-			switch (field_0x138) {
-			case 0:
-				break;
-			case 1:
-				JStage::TObject* object = NULL;
-				u32 v1 = -1;
-				if (v0 != 0) {
-					object = field_0x13c;
-					v1     = field_0x140;
-				}
-				pJSGObject_->JSGSetParent(object, v1);
-				break;
-			}
-		}
-		break;
-	}
+        bool v0 = (*(u32*)pContent != 0);
+        if (field_0x144 != v0) {
+            field_0x144 = v0;
+            switch (field_0x138) {
+            case 0:
+                break;
+            case 1:
+                JStage::TObject* object = NULL;
+                u32 v1 = -1;
+                if (v0 != 0) {
+                    object = field_0x13c;
+                    v1     = field_0x140;
+                }
+                pJSGObject_->JSGSetParent(object, v1);
+                break;
+            }
+        }
+        break;
+    }
 }
 
 /* 8028ABF0-8028AC04 285530 0014+00 1/0 0/0 0/0 .text
@@ -503,11 +233,11 @@ void JStudio_JStage::TAdaptor_actor::adaptor_do_PARENT_FUNCTION(
 void JStudio_JStage::TAdaptor_actor::adaptor_do_RELATION(
     JStudio::data::TEOperationData operation, void const* pContent, u32 uSize) {
     switch (operation) {
-	case JStudio::data::UNK_0x18:
+    case JStudio::data::UNK_0x18:
         JUT_ASSERT(324, pContent!=0);
         JUT_ASSERT(325, uSize!=0);
-		field_0x148 = adaptor_object_findJSGObject_((const char*)pContent);
-		break;
+        field_0x148 = adaptor_object_findJSGObject_((const char*)pContent);
+        break;
     }
 }
 
@@ -517,17 +247,17 @@ void JStudio_JStage::TAdaptor_actor::adaptor_do_RELATION(
 void JStudio_JStage::TAdaptor_actor::adaptor_do_RELATION_NODE(
     JStudio::data::TEOperationData operation, void const* pContent, u32 uSize) {
     switch (operation) {
-	case JStudio::data::UNK_0x18:
+    case JStudio::data::UNK_0x18:
         JUT_ASSERT(347, pContent!=0);
         JUT_ASSERT(348, uSize!=0);
-		field_0x14c = adaptor_object_findJSGObjectNode_(field_0x13c, (const char*)pContent);
-		break;
-	case JStudio::data::UNK_0x19:
+        field_0x14c = adaptor_object_findJSGObjectNode_(field_0x13c, (const char*)pContent);
+        break;
+    case JStudio::data::UNK_0x19:
         JUT_ASSERT(358, uSize==4);
         JUT_ASSERT(359, pContent!=0);
-		field_0x14c = *(u32*)pContent;
-		break;
-	}
+        field_0x14c = *(u32*)pContent;
+        break;
+    }
 }
 
 /* 8028ACA4-8028ACFC 2855E4 0058+00 1/0 0/0 0/0 .text
@@ -551,10 +281,10 @@ void JStudio_JStage::TAdaptor_actor::setJSG_ID_(void (JStage::TActor::*function)
                                                     JStudio::data::TEOperationData operation,
                                                     void const* pContent, u32 uSize) {
     switch (operation) {
-	case JStudio::data::UNK_0x19:
-		(((JStage::TActor*)pJSGObject_)->*function)(*(u32*)pContent);
-		break;
-	}
+    case JStudio::data::UNK_0x19:
+        (((JStage::TActor*)pJSGObject_)->*function)(*(u32*)pContent);
+        break;
+    }
 }
 
 /* 8028AD38-8028AECC 285678 0194+00 1/1 0/0 0/0 .text
@@ -647,77 +377,47 @@ void JStudio_JStage::TAdaptor_actor::getJSG_SRT_(JStudio::TControl const* pContr
 void JStudio_JStage::TAdaptor_actor::TVVOutput_ANIMATION_FRAME_::operator()(
     f32 param_1, JStudio::TAdaptor* adaptor) const {
     JStage::TActor* actor = static_cast<TAdaptor_actor*>(adaptor)->get_pJSG_();
-	// not sure what this bit is
-	u32 idx = *(u32*)(((u32)adaptor - 1) + field_0x8);
+    // not sure what this bit is
+    u32 idx = *(u32*)(((u32)adaptor - 1) + field_0x8);
 
-	u8 idx_lowBytes  = idx;
-	u8 idx_highBytes = idx >> 8;
+    u8 idx_lowBytes  = idx;
+    u8 idx_highBytes = idx >> 8;
 
-	f32 maxVal = (actor->*mMaxGetter)();
-	switch (idx_highBytes) {
-	case 1:
-		param_1 = maxVal - param_1;
-		break;
-	case 0:
-		break;
-	}
+    f32 maxVal = (actor->*mMaxGetter)();
+    switch (idx_highBytes) {
+    case 1:
+        param_1 = maxVal - param_1;
+        break;
+    case 0:
+        break;
+    }
 
-	if (maxVal > 0.0f) {
-		param_1 = (f32)(*JStudio::TFunctionValue::toFunction_outside(idx_lowBytes))(param_1, maxVal);
-	}
-	(actor->*mSetter)(param_1);
+    if (maxVal > 0.0f) {
+        param_1 = (f32)(*JStudio::TFunctionValue::toFunction_outside(idx_lowBytes))(param_1, maxVal);
+    }
+    (actor->*mSetter)(param_1);
 }
 
 /* 8028B138-8028B198 285A78 0060+00 2/1 0/0 0/0 .text
  * __dt__Q314JStudio_JStage14TAdaptor_actor26TVVOutput_ANIMATION_FRAME_Fv */
-// NONMATCHING
 JStudio_JStage::TAdaptor_actor::TVVOutput_ANIMATION_FRAME_::~TVVOutput_ANIMATION_FRAME_() {
 }
 
-/* ############################################################################################## */
-/* 803C5864-803C5888 022984 0010+14 2/2 0/0 0/0 .data
- * __vt__Q214JStudio_JStage81TVariableValueOutput_object_<Q214JStudio_JStage14TAdaptor_actor,Q26JStage6TActor>
- */
-SECTION_DATA extern void* data_803C5864[4 + 5 /* padding */] = {
-    (void*)NULL /* RTTI */,
-    (void*)NULL,
-    (void*)func_8028B5E0,
-    (void*)func_8028B580,
-    /* padding */
-    NULL,
-    NULL,
-    NULL,
-    NULL,
-    NULL,
+namespace JStudio_JStage {
+    
+/* 804311A0-804311E0 05DEC0 0040+00 2/3 0/0 0/0 .bss
+ * saoVVOutput___Q214JStudio_JStage14TAdaptor_actor             */
+const TAdaptor_actor::TVVOutputObject TAdaptor_actor::saoVVOutput_[] = {
+    TVVOutputObject(TAdaptor_actor::TEACTOR_1, &JStage::TActor::JSGSetAnimationTransition, &JStage::TActor::JSGGetAnimationTransition),
+    TVVOutputObject()
 };
 
-/* 8028B198-8028B568 285AD8 03D0+00 0/0 1/0 0/0 .text            __sinit_object-actor_cpp */
-void func_8028B198() {
-    // NONMATCHING
-}
+/* 80431204-80431298 05DF24 0090+04 2/3 0/0 0/0 .bss
+ * saoVVOutput_ANIMATION_FRAME___Q214JStudio_JStage14TAdaptor_actor */
+const TAdaptor_actor::TVVOutput_ANIMATION_FRAME_ TAdaptor_actor::saoVVOutput_ANIMATION_FRAME_[] = {
+    TVVOutput_ANIMATION_FRAME_(0, 305, &JStage::TActor::JSGSetAnimationFrame, &JStage::TActor::JSGGetAnimationFrame, &JStage::TActor::JSGGetAnimationFrameMax),
+    TVVOutput_ANIMATION_FRAME_(2, 309, &JStage::TActor::JSGSetTextureAnimationFrame, &JStage::TActor::JSGGetTextureAnimationFrame, &JStage::TActor::JSGGetTextureAnimationFrameMax),
+    TVVOutput_ANIMATION_FRAME_()
+};
 
-#pragma push
-#pragma force_active on
-REGISTER_CTORS(0x8028B198, func_8028B198);
-#pragma pop
-
-/* 8028B568-8028B580 285EA8 0018+00 1/1 6/6 0/0 .text            __ct__Q27JStudio14TVariableValueFv
- */
-JStudio::TVariableValue::TVariableValue() {
-    // NONMATCHING
-}
-
-/* 8028B580-8028B5E0 285EC0 0060+00 2/1 0/0 0/0 .text
- * __dt__Q214JStudio_JStage81TVariableValueOutput_object_<Q214JStudio_JStage14TAdaptor_actor,Q26JStage6TActor>Fv
- */
-extern "C" void func_8028B580(void* _this) {
-    // NONMATCHING
-}
-
-/* 8028B5E0-8028B610 285F20 0030+00 1/0 0/0 0/0 .text
- * __cl__Q214JStudio_JStage81TVariableValueOutput_object_<Q214JStudio_JStage14TAdaptor_actor,Q26JStage6TActor>CFfPQ27JStudio8TAdaptor
- */
-extern "C" void func_8028B5E0(void* _this, f32 param_0,
-                                  JStudio::TAdaptor* param_1) /* const */ {
-    // NONMATCHING
-}
+};  // namespace JStudio_JStage
